+++ spice/cuda/multi_snn.h
@@ -11,7 +11,7 @@ template <typename Model>
 class multi_snn : public ::spice::snn<Model>
 {
 public:
-	multi_snn( spice::util::layout desc, float dt, int delay = 1 );
+	multi_snn( spice::util::layout desc, float dt, int_ delay = 1 );
 	multi_snn( spice::snn<Model> const & net );
 
 	void step( std::vector<int> * out_spikes = nullptr ) override;
