+++ spice/snn_info.h
@@ -1,10 +1,13 @@
 #pragma once
 
+#include <spice/util/stdint.h>
+
+
 namespace spice
 {
 struct snn_info
 {
-	int num_neurons = 0;
+	int_ num_neurons = 0;
 	// num_synapses
 	// delay
 	// ...
