+++ spice_test/util/circular_buffer.cpp
@@ -19,7 +19,7 @@ TEST( CircBuffer, Ctor )
 		ASSERT_EQ( x.size(), 10 );
 		ASSERT_EQ( x.end() - x.begin(), 10 );
 
-		for( int i = 0; i < 10; i++ ) ASSERT_EQ( x[i], 0 );
+		for( int_ i = 0; i < 10; i++ ) ASSERT_EQ( x[i], 0 );
 
 		for( auto j : x ) ASSERT_EQ( j, 0 );
 	}
