+++ spice_test/util/meta.cpp
@@ -8,7 +8,7 @@
 using namespace spice::util;
 
 
-using test_types = type_list<int, int>;
+using test_types = type_list<int, int_>;
 enum attr
 {
 	One,
@@ -19,14 +19,14 @@ enum attr
 TEST( Meta, ForEach )
 {
 	{
-		std::tuple<int, int> x( 1, 2 );
+		std::tuple<int, int_> x( 1, 2 );
 		for_each( x, []( auto & elem ) { elem *= 2; } );
 		ASSERT_EQ( std::get<0>( x ), 2 );
 		ASSERT_EQ( std::get<1>( x ), 4 );
 	}
 
 	{
-		std::tuple<int, int> x( 0, 1 );
+		std::tuple<int, int_> x( 0, 1 );
 		for_each_i( x, []( auto elem, auto i ) { ASSERT_EQ( elem, i ); } );
 	}
 }
@@ -34,7 +34,7 @@ TEST( Meta, ForEach )
 TEST( Meta, Map )
 {
 	{
-		std::tuple<int, int> x( 1, 3 );
+		std::tuple<int, int_> x( 1, 3 );
 		auto y = map( x, []( auto elem ) { return elem * 1.5f; } );
 		ASSERT_EQ( std::get<0>( y ), 1.5f );
 		ASSERT_EQ( std::get<1>( y ), 4.5f );
